requirements_collect_data.txt
python-jobspy
pandas